module_17_5/requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
pydantic
python-slugify
